✨feat: feat : 쇼핑몰 판매 물품 등록, 판매 물품 조회, 판매 물품 재고 수정, 판매 물품 내역 조회. 총 4가지 기능 구현(이미지 제외)

+++ src/main/java/study/musinsa_project/entity/CartItems.java
@@ -9,9 +9,10 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class CartItems {
+public class CartItems
+{
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private int quantity;
